rpi-controller/requirements.txt
RPi.GPIO==0.7.0
paho-mqtt==2.1.0